Store VGA color attribute pre-shifted to eliminate repeated casts
Color is now a u16 initialized to 0x0700 (already in the high byte),
so writes are just `color | ch` instead of `static_cast<u16>(color) << 8 | ch`.

+++ include/hardware/terminal.hpp
@@ -66,7 +66,7 @@ private:
     u16* buffer;
     u8 x;
     u8 y;
-    u8 color;
+    u16 color;
 
     Port<u8> crtc_index;
     Port<u8> crtc_data;
